requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.7
django-bootstrap5==24.2
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==26.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
selenium==4.23.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
websocket-client==1.8.0
wsproto==1.2.0
